xc refactor : move subroutines around; add comment
