+++ source/module_xc/CMakeLists.txt
@@ -1,7 +1,10 @@
 list(APPEND objects
     xc_functional.cpp
     xc_functional_vxc.cpp
-    xc_gga_pw.cpp
+    xc_functional_gradcorr.cpp
+    xc_functional_wrapper_xc.cpp
+    xc_functional_wrapper_gcxc.cpp
+    xc_functional_wrapper_tauxc.cpp
     xc_funct_exch_lda.cpp
     xc_funct_corr_lda.cpp
     xc_funct_exch_gga.cpp
